requirements.txt
Flask>=2.3
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
sqlalchemy-serializer>=1.4
